+++ services/ske/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.11.0 (2024-03-28)
+
+- **Feature**: Waiters for async operation `StartCredentialsRotationWaitHandler` and `CompleteCredentialsRotationWaitHandler`
+
 ## v0.10.1 (2024-02-28)
 
 - Update `core` to [`v0.10.0`](../../core/CHANGELOG.md#v0100-2024-02-27)